+++ src/extension.ts
@@ -44,12 +44,16 @@ interface ProgressParams {
 
 export async function activate(context: ExtensionContext) {
   context.subscriptions.push(configureLanguage());
+  context.subscriptions.push(...registerCommands());
 
-  workspace.onDidOpenTextDocument(doc => whenOpeningTextDocument(doc, context));
-  workspace.textDocuments.forEach(doc => whenOpeningTextDocument(doc, context));
-  workspace.onDidChangeWorkspaceFolders(e =>
-    whenChangingWorkspaceFolders(e, context),
+  workspace.onDidOpenTextDocument(doc => whenOpeningTextDocument(doc));
+  workspace.onDidChangeWorkspaceFolders(e => whenChangingWorkspaceFolders(e));
+  window.onDidChangeActiveTextEditor(
+    ed => ed && whenOpeningTextDocument(ed.document),
   );
+  // Installed listeners don't fire immediately for already opened files, so
+  // trigger an open event manually to fire up RLS instances where needed
+  workspace.textDocuments.forEach(whenOpeningTextDocument);
 }
 
 export async function deactivate() {
@@ -57,10 +61,7 @@ export async function deactivate() {
 }
 
 // Taken from https://github.com/Microsoft/vscode-extension-samples/blob/master/lsp-multi-server-sample/client/src/extension.ts
-function whenOpeningTextDocument(
-  document: TextDocument,
-  context: ExtensionContext,
-) {
+function whenOpeningTextDocument(document: TextDocument) {
   if (document.languageId !== 'rust' && document.languageId !== 'toml') {
     return;
   }
@@ -96,7 +97,7 @@ function whenOpeningTextDocument(
     const workspace = new ClientWorkspace(folder);
     activeWorkspace = workspace;
     workspaces.set(folderPath, workspace);
-    workspace.start(context);
+    workspace.start();
   } else {
     const ws = workspaces.get(folderPath);
     activeWorkspace = typeof ws === 'undefined' ? null : ws;
@@ -140,10 +141,7 @@ function getOuterMostWorkspaceFolder(folder: WorkspaceFolder): WorkspaceFolder {
   return folder;
 }
 
-function whenChangingWorkspaceFolders(
-  e: WorkspaceFoldersChangeEvent,
-  context: ExtensionContext,
-) {
+function whenChangingWorkspaceFolders(e: WorkspaceFoldersChangeEvent) {
   _sortedWorkspaceFolders = undefined;
 
   // If a VSCode workspace has been added, check to see if it is part of an existing one, and
@@ -157,7 +155,7 @@ function whenChangingWorkspaceFolders(
       if (f === 'Cargo.toml') {
         const workspace = new ClientWorkspace(folder);
         workspaces.set(folder.uri.toString(), workspace);
-        workspace.start(context);
+        workspace.start();
         break;
       }
     }
@@ -175,8 +173,6 @@ function whenChangingWorkspaceFolders(
 
 // Don't use URI as it's unreliable the same path might not become the same URI.
 const workspaces: Map<string, ClientWorkspace> = new Map();
-let activeWorkspace: ClientWorkspace | null;
-let commandsRegistered: boolean = false;
 
 // We run one RLS and one corresponding language client per workspace folder
 // (VSCode workspace, not Cargo workspace). This class contains all the per-client
@@ -195,7 +191,7 @@ class ClientWorkspace {
     this.disposables = [];
   }
 
-  public async start(context: ExtensionContext) {
+  public async start() {
     if (!this.config.multiProjectEnabled) {
       warnOnMissingCargoToml();
     }
@@ -213,13 +209,9 @@ class ClientWorkspace {
     const isWin = process.platform === 'win32';
     const windowsHack = isWin ? '**' : '';
 
-    const pattern = this.config.multiProjectEnabled
-      ? `${windowsHack}${this.folder.uri.path}/**`
-      : undefined;
+    const pattern = `${windowsHack}${this.folder.uri.path}/**`;
 
-    const collectionName = this.config.multiProjectEnabled
-      ? `rust ${this.folder.uri.toString()}`
-      : 'rust';
+    const collectionName = `rust (${this.folder.uri.toString()})`;
     const clientOptions: LanguageClientOptions = {
       // Register the server for Rust files
 
@@ -263,12 +255,9 @@ class ClientWorkspace {
       clientOptions,
     );
 
-    const selector = this.config.multiProjectEnabled
-      ? { language: 'rust', scheme: 'file', pattern }
-      : { language: 'rust' };
+    const selector = { language: 'rust', scheme: 'file', pattern };
 
     this.setupProgressCounter();
-    this.registerCommands(context, this.config.multiProjectEnabled);
     this.disposables.push(activateTaskProvider(this.folder));
     this.disposables.push(this.lc.start());
     this.disposables.push(
@@ -287,47 +276,19 @@ class ClientWorkspace {
     }
 
     this.disposables.forEach(d => d.dispose());
-    commandsRegistered = false;
   }
 
-  private registerCommands(
-    context: ExtensionContext,
-    multiProjectEnabled: boolean,
-  ) {
-    if (!this.lc) {
-      return;
-    }
-    if (multiProjectEnabled && commandsRegistered) {
-      return;
-    }
+  public async restart() {
+    await this.stop();
+    return this.start();
+  }
 
-    commandsRegistered = true;
-    const rustupUpdateDisposable = commands.registerCommand(
-      'rls.update',
-      () => {
-        const ws =
-          multiProjectEnabled && activeWorkspace ? activeWorkspace : this;
-        return rustupUpdate(ws.config.rustupConfig());
-      },
-    );
-    this.disposables.push(rustupUpdateDisposable);
+  public runRlsCommand(cmd: Execution) {
+    return runRlsCommand(this.folder, cmd);
+  }
 
-    const restartServer = commands.registerCommand('rls.restart', async () => {
-      const ws =
-        multiProjectEnabled && activeWorkspace ? activeWorkspace : this;
-      await ws.stop();
-      commandsRegistered = true;
-      return ws.start(context);
-    });
-    this.disposables.push(restartServer);
-
-    this.disposables.push(
-      commands.registerCommand('rls.run', (cmd: Execution) => {
-        const ws =
-          multiProjectEnabled && activeWorkspace ? activeWorkspace : this;
-        runRlsCommand(ws.folder, cmd);
-      }),
-    );
+  public rustupUpdate() {
+    return rustupUpdate(this.config.rustupConfig());
   }
 
   private async setupProgressCounter() {
@@ -497,6 +458,34 @@ async function warnOnMissingCargoToml() {
       'A Cargo.toml file must be at the root of the workspace in order to support all features. Alternatively set rust-client.enableMultiProjectSetup=true in settings.',
     );
   }
+}
+
+/**
+ * Tracks the most current VSCode workspace as opened by the user. Used by the
+ * commands to know in which workspace these should be executed.
+ */
+let activeWorkspace: ClientWorkspace | null;
+
+/**
+ * Registers the VSCode [commands] used by the extension.
+ *
+ * [commands]: https://code.visualstudio.com/api/extension-guides/command
+ */
+function registerCommands(): Disposable[] {
+  return [
+    commands.registerCommand(
+      'rls.update',
+      () => activeWorkspace && activeWorkspace.rustupUpdate(),
+    ),
+    commands.registerCommand(
+      'rls.restart',
+      async () => activeWorkspace && activeWorkspace.restart(),
+    ),
+    commands.registerCommand(
+      'rls.run',
+      (cmd: Execution) => activeWorkspace && activeWorkspace.runRlsCommand(cmd),
+    ),
+  ];
 }
 
 /**